demo page refactoring
